IV Estimation of 2^K Factorial Design

Estimates principal stratum-specific effects and interactions in a 2^K factorial experiment

iv_factorial(formula, data, subset, method = "lm", level = 0.95)

Arguments

formula
formula specification of the factorial design with noncompliance. The right-hand side of the formula should have two components separated by the | symbol, with the first component containing the K binary treatment variables and the second component containing the K binary instruments associated with each treatment variable. The order of the variables in the formula must match.

data
A data.frame on which to apply the formula.

subset
subset of the data to pass to estimation.

method
character indiciating if the estimator should be "lm" using the least squares approach (default) or "cmd" to estimate via efficent minimum distance estimator.

level
the confidence level required.

Value

A list of class iv_factorial that contains the following components:

rho

vector of estimated compliance class probabilities.

psi

vector of the estimated conditional mean of the outcome within the compliance classes.

vcov

estimated asymptotic variance matrix of the combined rho and psi parameters.

pcafe_est

vector of estimated main effects of each factor among perfect compliers.

pcafe_se

vector of estimated standard errors for the estimated effects in tau.

pcafe_cis

a matrix of confidence intervals for the PCAFE estimates.

level

the confidence level of the returned confience intervals.

Details

This function estimates treatment effects for 2^K factorial experiments in the face of noncompliance on all factors. A monotonicity assumption is assumed for both treatment-instrument pairs, along with treatment exclusion. See Blackwell (2017) for more details on those assumptions.

The procedure uses iterative generalized method of moments (GMM) to estimate both the proportions of each compliance class (also known as principal strata) and the average potential outcomes within those classes. It also provides estimates of several one-way, joint, and interactive treatment effects within these classes.

Under the above assumptions, the compliance classes are the product of the compliance classes for each treatment-instrument pair. For instance, "cc" is the class that would comply with both treatments, "ca" is the class that would comply with the first treatment and always take the second treatment, and "cn" is the class that would comply with the first treatment and never take the second treatment. Finally, note that treatment effects are only well-defined for compliance classes for which there is compliance on at least one treatment.

References

Matthew Blackwell (2017) Instrumental Variable Methods for Conditional Effects and Causal Interaction in Voter Mobilization Experiments, Journal of the American Statistical Association, 112:518, 590-599, https://doi.org/10.1080/01621459.2016.1246363

Matthew Blackwell and Nicole Pashley (2020) "Noncompliance in Factorial Experiments." Working paper.

Author

Matt Blackwell

Examples

data(newhaven)

out <- iv_factorial(turnout_98 ~ inperson + phone | inperson_rand
  + phone_rand, data = newhaven)

summary(out)

#> 
#>  Main effects among perfect compliers:
#>                                      tval  pval
#> inperson        0.08204  0.35517  0.23099 0.817
#> phone           0.09320  0.36317  0.25664 0.797
#> inperson:phone -0.10452  0.29398 -0.35553 0.722
#> 
#> Estimated prob. of perfect compliers:  0.07969 	SE =  0.02311