Tidy an iv_finite_factorial object

Tidy summarizes information about the components of a model.

# S3 method for iv_finite_factorial
tidy(x, conf.level = 0.95, ...)

Arguments

x
An iv_factorial object produced by a call to iv_finite_factorial()

conf.level
The confidence level to use for the confidence interval. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments. Not used. Needed to match generic signature only.

Value

A tibble::tibble() with columns:

term

The name of the effect term.

estimand

Which complier effect being estimated.

estimate

The estimated value of the effect.

ci_1_lower

Lower bound for the first interval of the Fieller confidence regresion for the estimate.

ci_1_upper

Upper bound for the first interval of the Fieller confidence regresion for tshe estimate.

ci_2_lower

Lower bound for the second interval of the Fieller confidence regresion for the estimate. Only non-NA when the confidence region is disjoint.

ci_2_upper

Upper bound for the second interval of the Fieller confidence regresion for the estimate. Only non-NA when the confidence region is disjoint.

Author

Matt Blackwell